+++ front/lib/api/skills/detection/suggest_mcp_servers.ts
